+++ include/casm/clexmonte/misc/subparse_from_file.hh
@@ -33,24 +33,53 @@ template <typename RequiredType, typename T, typename... Args>
 std::shared_ptr<InputParser<RequiredType>> subparse_from_file(
     InputParser<T> &parser, fs::path option,
     std::vector<fs::path> search_path = {}, Args &&...args) {
+  // Logging for debug verbosity only
+  // (except for errors and warnings noted below)
+  auto &log = CASM::log();
+  log.begin_section<Log::verbose>();
+  log.increase_indent();
+  log.indent() << "- Attempting to subparse from file..." << std::endl;
+
+  jsonParser null_json;
+  auto null_subparser = std::make_shared<InputParser<RequiredType>>(
+      null_json, std::forward<Args>(args)...);
+
+  auto it = parser.self.find_at(option);
+  if (it == parser.self.end()) {
+    std::stringstream msg;
+    msg << "Error: missing required option '" << option.string() << "'.";
+    parser.insert_error(option, msg.str());
+
+    return null_subparser;
+  }
+
   std::string filepath;
   parser.require(filepath, option);
+  log.indent() << "- filepath: " << filepath << std::endl;
 
   fs::path resolved_path = resolve_path(filepath, search_path);
+  log.indent() << "- resolved_path: " << resolved_path << std::endl;
 
   if (!fs::exists(resolved_path)) {
+    log.indent() << "- resolved_path not found" << std::endl;
+    log.decrease_indent();
+    log.end_section();
+
     parser.insert_error(option, "Error: file not found.");
-    jsonParser json;
-    auto subparser = std::make_shared<InputParser<RequiredType>>(
-        json, std::forward<Args>(args)...);
-    return subparser;
+    return null_subparser;
   }
+
+  log.indent() << "- Reading file..." << std::endl;
   jsonParser json{resolved_path};
+  log.indent() << "- Subparsing..." << std::endl;
   auto subparser = std::make_shared<InputParser<RequiredType>>(
       json, std::forward<Args>(args)...);
 
-  auto &log = CASM::log();
   if (!subparser->valid()) {
+    log.indent() << "- Subparsing: failed" << std::endl << std::endl;
+
+    // Always log this:
+    log.begin_section<Log::none>();
     log << std::endl;
     log << "~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~" << std::endl;
     log << "~~~  Failed parsing input file  ~~~" << std::endl;
@@ -65,8 +94,9 @@ std::shared_ptr<InputParser<RequiredType>> subparse_from_file(
     }
     if (json.is_obj()) {
       jsonParser report = make_report(*subparser);
-      log.indent() << report << std::endl << std::endl;
+      log << report << std::endl << std::endl;
     }
+    log.end_section();
 
     parser.insert_error(
         option, "Error: Failed to parse file: " + resolved_path.string());
@@ -81,8 +111,13 @@ std::shared_ptr<InputParser<RequiredType>> subparse_from_file(
                               "@(/" + warning.first.string() + "): " + msg);
       }
     }
+  } else {
+    log.indent() << "- Subparsing: succeeded" << std::endl;
   }
+
   if (subparser->all_warnings().size()) {
+    // Always log this:
+    log.begin_section<Log::none>();
     log << std::endl;
     log << "~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~" << std::endl;
     log << "~~~ Warnings parsing input file ~~~" << std::endl;
@@ -95,6 +130,7 @@ std::shared_ptr<InputParser<RequiredType>> subparse_from_file(
       jsonParser report = make_report(*subparser);
       log.indent() << report << std::endl << std::endl;
     }
+    log.end_section();
 
     parser.insert_warning(
         option, "Warning: warnings for file: " + resolved_path.string());
@@ -105,6 +141,10 @@ std::shared_ptr<InputParser<RequiredType>> subparse_from_file(
       }
     }
   }
+
+  log.indent() << "- Subparsing finished..." << std::endl;
+  log.end_section();
+  log.decrease_indent();
 
   subparser->type_name = CASM::type_name<RequiredType>();
   return subparser;